% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcgis_addressfinder.R,
%   R/arcgis_addressfinder_httr.R
\name{addressFinder}
\alias{addressFinder}
\title{Call Arcgis API "geocodeAddresses" operation for bulk address geocoding.}
\usage{
addressFinder(address_table, url)

addressFinder(address_table, url)
}
\arguments{
\item{address_table}{A table with addresses to be geocoded. Columns used in geocoding include required:"address" optional:"address2" "address3" "city" "postal" "countryCode"}

\item{url}{An URL with which to generate your Arcgis access token "https://www.arcgis.com/sharing/oauth2/token?client_id=<YOUR CLIENT ID>&grant_type=client_credentials&client_secret=<YOUR CLIENT SECRET>&f=pjson" see https://developers.arcgis.com/rest/geocode/api-reference/geocoding-authenticate-a-request.htm}
}
\value{
A table of Geocoded results of matched addresses merged onto your origional columns.

A table of Geocoded results of matched addresses merged onto your origional columns.
}
\description{
Call Arcgis API "geocodeAddresses" operation for bulk address geocoding.

Call Arcgis API "geocodeAddresses" operation for bulk address geocoding.
}
